all functions in wordsearch can't test (?) need to clean code
